添加了关于界面

+++ UI/widget/RTDDeviceInfoTableModel.cpp
@@ -29,8 +29,8 @@ bool RTDDeviceInfoRTableModel::setHeaders(const QVector<QString>& DeviceList, co
 		this->_dataList[i].resize(DeviceSensors.size());
 	}
 
-	emit this->headerDataChanged(Qt::Horizontal, 0, std::min(this->_header.size() - 1, static_cast<long long>(0)));
-	emit this->headerDataChanged(Qt::Vertical, 0, std::min(this->_DeviceName.size() - 1, static_cast<long long>(0)));
+	emit this->headerDataChanged(Qt::Horizontal, 0, std::min(static_cast<long long>(this->_header.size() - 1), static_cast<long long>(0)));
+	emit this->headerDataChanged(Qt::Vertical, 0, std::min(static_cast<long long>(this->_DeviceName.size() - 1), static_cast<long long>(0)));
 	return true;
 }
 
@@ -45,7 +45,7 @@ bool RTDDeviceInfoRTableModel::setHeaders(const QVector<QString>& UserDataHeader
 	this->_header = UserDataHeader.toList();
 	this->_dataList.resize(1);
 	this->_dataList[0].resize(UserDataHeader.size());
-	emit this->headerDataChanged(Qt::Horizontal, 0, std::min(this->_header.size() - 1, static_cast<long long>(0)));
+	emit this->headerDataChanged(Qt::Horizontal, 0, std::min(static_cast<long long>(this->_header.size() - 1), static_cast<long long>(0)));
 	emit this->headerDataChanged(Qt::Vertical, 0, 0);
 	return true;
 }
